Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.21.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.0.1
dynaconf